Added 'Keybinds' mod intended to replace a few other mods that only sets keybinds.
This commit is dependent on a pull request #1 in the Input repo.

The mods it replaces are:
ObjectDumperMod, CXXHeaderGeneratorMod and UHTCompatibleHeaderGeneratorMod.

It's backwards compatible and if a user has one of those mods, the key will still be controlled from that mod instead of this new mod.
This is so that it won't cause problems for users that have already modified their key bindings.
This should be mentioned in the changelog.