Halved the number of objects per chunk in the live view.
I also renamed "chunks" to "groups" to distance this mechanic from the chunks system in UE.
While the two systems are similar, they are not the same and the live view isn't tied to the UE chunks system.

Also added 'LiveViewObjectsPerGroup' to the settings file to allow the user to control this number in case they experience lag.