CppMods are now loaded earlier
This is achieved by splitting UE4SSProgram constructor into an init
function, and only initializing the bare minimum in the constructor.
This allows for the constructor to be in DllMain and run our code before
any game code actually runs.

CppMods also got a new `dll_load` event, which notifies them when any
dll got loaded into the process, this allows for mods to monitor when a
dll that they're interested in loaded, and act accordingly.