new proxy injection way
This patch removes the need to compile ue4ss inside of a proxy. Instead,
this compiles the proxy separately, which then loads ue4ss.dll. This
allows for more flexibility when rebuilding the proxy as a different
dll.

Because the proxy uses codegen, we can swap to another dll by changing
the `UE4SS_PROXY_PATH` variable to point to a new dll.