Added UEHelpers.FindOrAddFName
This function will return the FName if it exists, and otherwise it will create it.

The UE documentation for 'EFindName::FNAME_Add' doesn't seem to be correct.
It says: "Find a name or add it if it doesn't exist.", however, if the name already exists, it will add a new name with a number suffixed to it like "_0".
This new UEHelpers function doesn't have this functionality and instead will return the original name without any extra suffix.

Also updated the docs to correctly point out that FNAME_Add is the default for the FName constructor, not FNAME_Find.
Also added EFindName to the docs.