The original return value is now passed to the second RegisterHook callback
If there's a second callback given to 'RegisterHook', the return value passed to that callback as the second param is now always the original return value, even if you made modifications to it in the first callback by returning something.

This was accomplished by leaving the return value of the first callback on the Lua stack until the second callback has been called and then both return values are processed.