Fixed BPModLoaderMod giving "conversion error"
This happened when a directory name contained a unicode character.

According to my tests with this fix applied, unicode characters translate properly between wstring and our own to_string function and iterating the results of "IterateGameDirectories" and using Luas "print" function will properly display the unicode characters in the console.

Added two notes to the changelog, one documenting the underlying problem that was fixed, and another one documenting the upstream feature that was fixed because of that.
This is so that people that experience this problem can recognize it as fixed in the changelog.