Fixed keybinds created from C++ mods executing multiple times This only happened if the hot-reload feature was used to reload all mods.
This was fixed by introducing CppUserModBase::register_keydown_event.
It's identical to the one in UE4SSProgram, except it keeps track of which C++ mod registered the event so that it can be "unregistered" when the mod is reloaded.

Mods that continue to use the UE4SSProgram version will continue to have the same problem.

Updated changelog