Replaced usage of 'SetDefaultDllDirectories'
This was done to fix PalServer not working properly after an update.
The usage of 'SetDefaultDllDirectories' was there to allow C++ mods that have dynamic dependencies to store those dependencies right next to main.dll instead of in the Win64 directory.

This behavior remains intact, but we're using 'LoadLibraryEx' instead of 'LoadLibrary' to set the path specifically for main.dll instead of globally with 'SetDefaultDllDirectories'.

See #452 for discussion around this problem.