Refactor File APIs, using UTF-8 as native file encoding Define prupose driven string types in Macros.hpp
Use IOSTR/File::StringType for IO purpose, it's pinned to utf8/string type
Use SYSSTR/SystemStringType for interacting with STL, it's the prefer type
based on platform; On windows this is wchar_t/wstring and on Linux
this is char/string
Use STR/UEStringType for interacting with Unreal Engine, it's utf16.
On Windows we uses wchar_t/wstring to avoid extra copy/casting from u16string
On Linux we uses u16string which cannot be used in many STL functions.

With this patch, the only changes on string type on Windows should just be the
type aliases, and almost all actual types should stay unchanged.

For the IO, write_string_to_file will keep receiving utf16, aka wchar_t on windows.
A new write_file_string_to_file is added to allow receiving of utf8 string

A new WinFile using WinAPI is provided and to avoid the buggy BOM detection on the old code.
It also avoids the use of wstring to handle UTF-8 chars.