GetPlayer != GetPawn
This is misleading and incorrect. "Player" in the context of UE is a completely different to a pawn object. You always have a player on a client, be it `LocalPlayer` or `NetClient`, but there are plenty of games that never use Pawn, nor for PlayerController nor for any of the other entities.
This should be removed in major release or rewritten to return actual player Object